+++ data-raw/ReloadParameters.R
@@ -10,10 +10,10 @@ SpParamsDefinition <-as.data.frame(readxl::read_xlsx("data-raw/SpParamsDefinitio
 SpParamsDefinition$Definition = stringi::stri_enc_toascii(SpParamsDefinition$Definition)
 SpParamsDefinition$Units = stringi::stri_enc_toascii(SpParamsDefinition$Units)
 usethis::use_data(SpParamsDefinition, overwrite = T)
-SpParamsUS <-as.data.frame(readxl::read_xlsx("data-raw/SpParamsUS.xlsx",
-                                             sheet="SpParamsUS", na = "NA"), stringsAsFactors=FALSE)
-usethis::use_data(SpParamsUS, overwrite = T)
-rm(SpParamsUS)
+# SpParamsUS <-as.data.frame(readxl::read_xlsx("data-raw/SpParamsUS.xlsx",
+#                                              sheet="SpParamsUS", na = "NA"), stringsAsFactors=FALSE)
+# usethis::use_data(SpParamsUS, overwrite = T)
+# rm(SpParamsUS)
 
 
 # Initial table